% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarabc.R
\name{sim_sar}
\alias{sar_sim}
\alias{sim_sar}
\title{Simulates a SAR model}
\usage{
sim_sar(G, X, rho, beta, p = 5, xmean = rep(0, nrow(G)))

sar_sim(G, X, rho, beta, p = 5, xmean = rep(0, nrow(G)))
}
\arguments{
\item{G}{Square matrix of class \code{dgCMatrix} of size \eqn{n\times n}{n*n}.}

\item{X}{Numeric matrix of covariates of size \eqn{n\times k}{n * k}.}

\item{rho}{Numeric scalar.}

\item{beta}{Numeric colvector of length \eqn{k}.}

\item{p}{Number of iterations in the approximation to \eqn{(I_n - \rho W)^{-1}}.}

\item{xmean}{Numeric vector of length \eqn{n}.}
}
\value{
A numeric vector of length \eqn{n}.
}
\description{
Simulates a SAR model
}
\details{
Follows a model SAR

\deqn{%
y = \rho \mathbf{W}y + X\beta + \varepsilon,\quad\varepsilon\sim \mbox{MVN}(0,\sigma^2 I_n)
}{%
y = rho * W * y + X*beta + eps, eps ~ MVN(0, sigma^2 *I_n)
}
}

